Moved uuid object to new file uuidObject.ts, edited changefile verbage
