+++ packages/teams-js/src/internal/interfaces.ts
@@ -1,5 +1,3 @@
-import { generateGUID, validateUuid } from './utils';
-
 /**
  * @hidden
  * Hide from docs
@@ -51,20 +49,4 @@ export interface ExtendedWindow extends Window {
 export interface DOMMessageEvent {
   func: string;
   args?: any[];
-}
-
-/**
- * @internal
- * Limited to Microsoft-internal use
- *
- * UUID object
- */
-export class UUID {
-  public constructor(private readonly uuid: string = generateGUID()) {
-    validateUuid(uuid);
-  }
-
-  public toString(): string {
-    return this.uuid;
-  }
 }
