+++ packages/teams-js/src/internal/uuidObject.ts
@@ -1,0 +1,17 @@
+import { generateGUID, validateUuid } from './utils';
+
+/**
+ * @internal
+ * Limited to Microsoft-internal use
+ *
+ * UUID object
+ */
+export class UUID {
+  public constructor(private readonly uuid: string = generateGUID()) {
+    validateUuid(uuid);
+  }
+
+  public toString(): string {
+    return this.uuid;
+  }
+}
